arch/arm, kernel/sched: move yield functionality into arch for optimization
When CONFIG_SCHED_YIELD_OPTIMIZATION is enabled, it does not use
kernel function so that it does not need to do separately.
As others do like up_unblock_task, up_reprioritize_rtr, let's
do entire operation in arch code.
