+++ os/arch/arm/src/armv7-m/up_schedyield.c
@@ -22,13 +22,17 @@
 
 #include <tinyara/config.h>
 
-#ifdef CONFIG_TASK_MONITOR
 #include <sys/types.h>
-#endif
+#include <debug.h>
+#include <queue.h>
 #ifdef CONFIG_SUPPORT_COMMON_BINARY
 #include <stdint.h>
 #endif
+
 #include <tinyara/sched.h>
+#if CONFIG_RR_INTERVAL > 0
+#include <tinyara/clock.h>
+#endif
 #ifdef CONFIG_ARMV7M_MPU
 #include <tinyara/mpu.h>
 #endif
@@ -69,80 +73,124 @@ extern uint32_t g_umm_app_id;
  *
  *   1) The priority of the currently running task is either equal or more than other ready to run task
  *
- * Inputs:
- *   tcb: The TCB of the task that has been reprioritized
- *
  ****************************************************************************/
 
-void up_schedyield(struct tcb_s *rtcb)
+void up_schedyield(void)
 {
-	if (current_regs) {
-		/* Yes, then we have to do things differently.
-		 * Just copy the current_regs into the OLD rtcb.
+	struct tcb_s *rtcb = this_task();
+	struct tcb_s *ntcb = rtcb->flink;
+
+	if (ntcb == NULL) {
+		sdbg("Fail to yield, there is no next tcb. Current task : %d\n", rtcb->pid);
+		return;
+	}
+
+	/* Yielding the CPU resource to other tasks is possible only if other tasks
+	 * priority is either equal or greater than currently running task.
+	 * It's achieved by rescheduling the current task behind any other tasks at
+	 * same priority
+	 */
+
+	if (rtcb->sched_priority <= ntcb->sched_priority) {
+
+		/* Remove the TCB from the ready-to-run list */
+
+		dq_rem((FAR dq_entry_t *)rtcb, (FAR dq_queue_t *)&g_readytorun);
+
+		/* Add the task in the correct location in the prioritized
+		 * g_readytorun task list
 		 */
 
-		up_savestate(rtcb->xcp.regs);
+		sched_addprioritized(rtcb, (FAR dq_queue_t *)&g_readytorun);
 
-		/* Restore the exception context of the rtcb at the (new) head
-		 * of the g_readytorun task list.
+		/* The current tcb was added in the middle of the ready-to-run list */
+
+		rtcb->task_state = TSTATE_TASK_READYTORUN;
+
+		/* we are going to do a context switch, then now it's the right
+		 * time to add any pending tasks back into the ready-to-run task
+		 * list now
 		 */
 
-		rtcb = this_task();
+		if (g_pendingtasks.head) {
+			sched_mergepending();
+		}
+
+		/* Get new head of the list */
+
+		ntcb = this_task();
+
+#if CONFIG_RR_INTERVAL > 0
+		ntcb->timeslice = MSEC2TICK(CONFIG_RR_INTERVAL);
+#endif
+
+		/* A context switch will occur. */
+
+		ntcb->task_state = TSTATE_TASK_RUNNING;
+
 #ifdef CONFIG_TASK_SCHED_HISTORY
 		/* Save the task name which will be scheduled */
 
-		save_task_scheduling_status(rtcb);
+		save_task_scheduling_status(ntcb);
 #endif
 
-		/* Restore the MPU registers in case we are switching to an application task */
-#ifdef CONFIG_ARMV7M_MPU
-		/* Condition check : Update MPU registers only if this is not a kernel thread. */
+		/* Are we in an interrupt handler? */
 
-		if ((rtcb->flags & TCB_FLAG_TTYPE_MASK) != TCB_FLAG_TTYPE_KERNEL) {
+		if (current_regs) {
+
+			/* Yes, then we have to do things differently.
+			 * Just copy the current_regs into the OLD rtcb.
+			 */
+
+			up_savestate(rtcb->xcp.regs);
+
+			/* Restore the exception context of the ntcb at the (new) head
+			 * of the g_readytorun task list.
+			 */
+
+			/* Restore the MPU registers in case we are switching to an application task */
+#ifdef CONFIG_ARMV7M_MPU
+			/* Condition check : Update MPU registers only if this is not a kernel thread. */
+
+			if ((ntcb->flags & TCB_FLAG_TTYPE_MASK) != TCB_FLAG_TTYPE_KERNEL) {
 #if defined(CONFIG_APP_BINARY_SEPARATION)
-			for (int i = 0; i < 3 * MPU_NUM_REGIONS; i += 3) {
-				up_mpu_set_register(&rtcb->mpu_regs[i]);
-			}
+				for (int i = 0; i < 3 * MPU_NUM_REGIONS; i += 3) {
+					up_mpu_set_register(&ntcb->mpu_regs[i]);
+				}
 #endif
 #ifdef CONFIG_MPU_STACK_OVERFLOW_PROTECTION
-			up_mpu_set_register(rtcb->stack_mpu_regs);
+				up_mpu_set_register(ntcb->stack_mpu_regs);
 #endif
-		}
+			}
 #endif
 
 #ifdef CONFIG_SUPPORT_COMMON_BINARY
-		if (g_umm_app_id) {
-			*g_umm_app_id = rtcb->app_id;
-		}
+			if (g_umm_app_id) {
+				*g_umm_app_id = ntcb->app_id;
+			}
 #endif
+
 #ifdef CONFIG_TASK_MONITOR
-		/* Update rtcb active flag for monitoring. */
+			/* Update rtcb active flag for monitoring. */
 
-		rtcb->is_active = true;
+			ntcb->is_active = true;
 #endif
-		/* Then switch contexts */
 
-		up_restorestate(rtcb->xcp.regs);
-	}
-	/* No, then we will need to perform the user context switch */
+			/* Then switch contexts */
 
-	else {
-		/* Switch context to the context of the task at the head of the
-		 * ready to run list.
-		 */
+			up_restorestate(ntcb->xcp.regs);
+		}
 
-		struct tcb_s *nexttcb = this_task();
-#ifdef CONFIG_TASK_SCHED_HISTORY
-		/* Save the task name which will be scheduled */
+		/* No, then we will need to perform the user context switch */
 
-		save_task_scheduling_status(nexttcb);
-#endif
-		up_switchcontext(rtcb->xcp.regs, nexttcb->xcp.regs);
+		else {
+			up_switchcontext(rtcb->xcp.regs, ntcb->xcp.regs);
 
-		/* up_switchcontext forces a context switch to the task at the
-		 * head of the ready-to-run list.  It does not 'return' in the
-		 * normal sense.  When it does return, it is because the blocked
-		 * task is again ready to run and has execution priority.
-		 */
+			/* up_switchcontext forces a context switch to the task at the
+			 * head of the ready-to-run list.  It does not 'return' in the
+			 * normal sense.  When it does return, it is because the blocked
+			 * task is again ready to run and has execution priority.
+			 */
+		}
 	}
 }
